requirements.txt
anthropic>=0.25.0
pillow>=10.0.0
python-dotenv>=1.0.0